requirements.txt
pact-python
requests
pytest
fastapi